WIP Adding CLI app
This is a WIP. We're looking into adding a CLI from which we can launch
a full node. We add a new app under the mana umbrella to put this code.

For now, we put some sample code in a mix task that uses some sample
data from a ropsten chain, and it validates and includes the blocks in
the block chain. This can be used with `mix mana`.

In the future, we would likely want this to be a full command line
interface where we can pass options like which chain to connect, etc.