Generate Client app and move sync task
We generate a new app called `Client`. This will be the main client/node
that we launch in `mana`. For now, we have moved the `sync` mix task
that lived in `ex_wire` for a little bit that has all the logic to
process blocks requested from a remote tree.