Subscribe Sync task to TCP as a server
The TCP module in ExWire has two ways to subscribe to it. One is by
giving it a function that will be called in our module (that was the
implementation we were following). The other is to expect a message from
it. The latter implementation makes more sense as a process subscribing
to receive messages. So we update the sync task to do so, and that
removes the need the have a separate function that simply sends
ourselves a message.