Create Client.Node (with sync mix task code)
We create a `Client.Node` module that uses a `GenServer`. The code that
we previously had in the `sync` mix task is now living in the GenServer.
We also supervise the node so that it restarts if it crashes.

We update the `mix sync` task to simply run `mix run --no-halt`.

Finally, we add a lot of `handle_info/2` functions to the `Client.Node`
to outline what types of messages we are receiving. For now we can
ignore most messages but we will want to use the information from some
of them in the future, and we'll likely have to respond to other
messages (if a node is trying to get info from our node) in order to
satisfy the specification.