WIP begin extraction of handle_auth into RLPx module
The `Handshake` module is doing a little too much in the
`try_handle_auth` function. I think we can move that into a higher level
module `RLPx` that is responsible for coordinating the crypto handshake
across the `Hanshake` and `Secrets` modules.

We extract that, but we need to improve tests and update the code that
currently calls `Handshake.try_handle_auth`.