Fix Handshake.try_handle_auth
Handshake.try_handle_auth had incorrect function signatures in that it
assumed we would have more information to provide to the function (like
the initiator node's id, but we get that from decoding the auth message
itself). It also called functions passing arguments incorrectly -
somehow this did not fail at compile time, and it did not have any test
coverage.